// types.ts
export type ViewName = 'dashboard' | 'residents' | 'medications' | 'medical-care' | 'technical-care' | 'demands' | 'inventory' | 'operations' | 'personal-items' | 'reports' | 'settings' | 'employees' | 'evolutions' | 'admin-panel';